feat(frontend): 状态卡片、指针、组件迁移及多项UI优化
- feat(ui): 更新“当前状态”Bento卡片，使用Emoji图标及含花体字符的描述。
- fix(pointer): 修正自定义指针在模态框中的定位、Emoji居中及原生光标覆盖。
- refactor(routing): 将main.tsx内联页面组件迁移到独立文件，解决Linter问题。
- fix(styling): 解决全局CSS导致的字体覆盖，并用overflow-y:scroll修复布局位移。
- docs: 更新design.md及To-do列表，加入动态样式化文本的设计思路。
