+++ backend/To-do.md
@@ -45,6 +45,10 @@
     -   [ ] 用户认证 API (`/login`, `/register`, `/me`)
     -   [ ] 博文 CRUD API (创建、更新、删除博文)
     -   [ ] 分类/标签管理 API
+    -   [ ] **(新增) 动态可样式化文本 API:**
+        -   [ ] 设计并实现数据结构（如JSON字段，包含文本片段和样式键）用于存储可自定义样式的文本内容（例如，首页"当前状态"卡片的描述）。
+        -   [ ] 提供 API 端点以供前端获取这些结构化文本数据。
+        -   [ ] 在管理后台的保存逻辑中，处理从富文本编辑器（如TipTap）生成的结构化数据并持久化。
 
 ## 其他
