+++ frontend/To-do&Completed.md
@@ -44,7 +44,7 @@
 
 ## 阶段三：页面开发 - 核心页面 (根据 `design.md`)
 
--   [ ] **Homepage (`/`)**
+-   [x] **Homepage (`/`)**
     -   [x] Hero Section (座右铭, 粒子背景) - *假设已完成，如未完成请取消勾选*
     -   [x] Bento Grid 布局实现 (集成 `bento-grid.tsx`)
         -   [x] 自我介绍格 (简介, 头像) - *部分内容已配置*
@@ -103,6 +103,9 @@
     -   [ ] 分类管理 (CRUD)
     -   [ ] 标签管理 (CRUD)
     -   [ ] (可选) 媒体管理
+    -   [ ] **(新增) 动态文本样式支持:** 
+        -   [ ] 前端实现对从后端获取的结构化文本（含样式键）的解析和渲染逻辑（例如，用于首页"当前状态"卡片描述）。
+        -   [ ] 在管理后台的编辑器中（如TipTap）支持用户选择文本并应用预定义的样式（如正常、斜体、自定义字体）。
 -   [ ] **个人工具 (Workbench - 占位符，根据 `design.md` 未来考虑)**
     -   [ ] RSS 阅读器 (订阅源管理, 文章列表, 阅读状态)
     -   [ ] 待办事项列表 (任务管理, 状态, 分类/优先级)
