+++ frontend/src/index.css
@@ -47,6 +47,15 @@
   font-display: swap;
 }
 
+/* Dancing Script - For the special 'L' */
+@font-face {
+  font-family: 'Dancing Script Custom'; /* Custom name to avoid conflicts if Google Font version was ever imported */
+  src: url('./assets/fonts/dancing_script/DancingScript-VariableFont_wght.woff2') format('woff2-variations');
+  font-weight: 400 700; /* Dancing Script variable font typically supports this range */
+  font-style: normal;
+  font-display: swap;
+}
+
 @theme inline {
   --radius-sm: calc(var(--radius) - 4px);
   --radius-md: calc(var(--radius) - 2px);
@@ -127,3 +136,17 @@
     @apply bg-background text-foreground;
   }
 }
+
+/* Custom base styles */
+html {
+  @apply bg-background; /* 确保 html 元素也使用主题背景色 */
+  min-height: 100%;
+  overflow-y: scroll;   /* 保持：防止滚动条显隐造成的布局位移 */
+}
+
+body {
+  min-height: 100%;     /* 保持：确保 body 也能撑满高度 */
+  /* 字体、文本颜色和 body 的背景色由 Tailwind 的 base 层和您在 :root 及 @layer base 中的主题设置处理 */
+}
+
+/* Further global styles can be added here */
