+++ frontend/src/pages/AboutPage.tsx
@@ -1,12 +1,12 @@
 // import React from 'react';
 import { Link } from 'react-router-dom';
 
-const BlogListPage = () => (
+const AboutPage = () => (
   <div className="container mx-auto p-4">
-    <h2 className="text-2xl font-bold mb-4">Blog List Page</h2>
-    <p>TODO: Implement blog list</p>
+    <h2 className="text-2xl font-bold mb-4">About Me</h2>
+    <p>TODO: Implement about me page content. This page will describe you.</p>
     <Link to="/" className="text-blue-500 hover:underline">Go to Home</Link>
   </div>
 );
 
-export default BlogListPage; 
+export default AboutPage;
