+++ frontend/src/pages/AcademicPage.tsx
@@ -1,12 +1,12 @@
-import React from 'react';
+// import React from 'react';
+import { Link } from 'react-router-dom';
 
-const AcademicPage: React.FC = () => {
-  return (
-    <div>
-      <h1>Academic Page Container</h1>
-      {/* Future content for the academic page will go here */}
-    </div>
-  );
-};
+const AcademicPage = () => (
+  <div className="container mx-auto p-4">
+    <h2 className="text-2xl font-bold mb-4">Academic Page</h2>
+    <p>TODO: Implement academic page content</p>
+    <Link to="/" className="text-blue-500 hover:underline">Go to Home</Link>
+  </div>
+);
 
 export default AcademicPage;
