+++ frontend/src/pages/AdminLoginPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const AdminLoginPage = () => (
+  <div className="container mx-auto p-4">
+    <h2 className="text-2xl font-bold mb-4">Admin Login Page</h2>
+    <p>TODO: Implement admin login page content</p>
+    <Link to="/" className="text-blue-500 hover:underline">Go to Home</Link>
+  </div>
+);
+
+export default AdminLoginPage;
