+++ frontend/src/pages/ProjectsPage.tsx
@@ -1,0 +1,12 @@
+// import React from 'react';
+import { Link } from 'react-router-dom';
+
+const ProjectsPage = () => (
+  <div className="container mx-auto p-4">
+    <h2 className="text-2xl font-bold mb-4">My Projects</h2>
+    <p>TODO: Implement projects page content. This page will showcase your projects.</p>
+    <Link to="/" className="text-blue-500 hover:underline">Go to Home</Link>
+  </div>
+);
+
+export default ProjectsPage;
